Added point size, point weight and labels for the format window. Point Size determines the size of the poitns, Point Weight determines the line width for the points being drawn. Does not do anything for fill shapes. Fixed the SVG clipPath element to go into a defs region so now svg files work better, and import into inkscape. Added a check with draw line so that the SVG painter does not inlcude lines that are outside of the clip region. Changed the graphPainter to use doubles in the signiture since svg can have double precision elements. PanelPainter now converts these elements.

+++ lightgraph/painters/GraphPainter.java
@@ -19,10 +19,10 @@ public interface GraphPainter {
     public void drawPath(Shape s);
     public void setColor(Color c);
     public Color getColor();
-    public void drawLine(int x0, int y0, int x1, int y1);
+    public void drawLine(double x0, double y0, double x1, double y1);
     public void fill(Shape s);
     public void setClip(int x, int y, int w, int h);
-    public void drawString(String s, int x, int y);
+    public void drawString(String s, double x, double y);
     public void setLineWidth(double f);
     public void restoreLineWidth();
 }
